replace float type

+++ schema.py
@@ -19,10 +19,10 @@ df_schema = StructType(
 )
 df_rounded_schema = StructType(
     [
-        StructField("timestamp", IntegerType(), True),
+        StructField("timestamp", FloatType(), True),
         StructField("content_id", IntegerType(), True),
         StructField("counter", IntegerType(), True),
-        StructField("timestamp_", IntegerType(), True)
+        StructField("timestamp_", FloatType(), True)
     ]
 )
